LEDModule Initialized
Decided on using Sk6812 Based LEDS as WS2812B are out of stock on JLCPCB.

Layed out all 15 LEDS on LED Module Schematic. Exported SVG of Edgecuts for PCB from Fusion360 and used Inkscape to convert them to a DXF file for use as edge cuts in the PCB Editor.